ProfileView: eradicate table views
The tableviews arent really necessary anymore.  We display 0-2 static
entries for each section, and then flip to an alternate view.  Remove
the table and create a simpler view that acts as the cell.  Once the
height issues are figured out, the main view itself should be
transitioned into a scrollview which would bring this inline with the
mocks.