ProfileView: split out configuration of the view
Refactor the supporter and events view into separate functions.  This
simplifies the configureView function, allows us to make it more dynamic
and to properly handle the rendering of the supporters and events in the
range [0,n].  It also opens up the possibility for refactoring further
into a custom container view to simplify the layout constraints.