WIP/Dashboard: setup code for the new dashboard UX
Rework the existing code to setup the dashboard style layout instead.
This effectively serves as a gut of the original application code.  Some
of the components will still be useful in the future for reference, so
leaving the code behind for now.  This sets up the core of the card view
layout and adds cards on the dashboard.  The profile and activity
cards have sufficiently enough mocked up to render something similar to
the desired UX.  The fundraising card is still incomplete and the
challenge card still needs to be fixed up to render the progress bars.

The UI elements still have an incorrect drop shadow on the cards and the
border of the cards needs to be adjusted.  The font styling is still
incorrect and some of the UI elements need to be adjusted still
(primarily the team name/label).  The status buttons in the profile card
is still missing.  The popup segue needs to be altered for the items or
fixed to dynamically size the popup (preferable).  The buttons are not
yet given touch targets for the callback.