Login: add in AKF Login interstitial
Make the user perform the AKF Login action if no AKF ID has been stored.
This sets up the hook for the JS callback, but that has not yet been
provisioned.  In the mean time, provide a button to dismiss the
interstitial dialog.  This uses WKWebView rather than a
SFSafariViewController as the latter does not allow for the JavaScript
bridging.