Team: list members of the team
Attempt to query and join the information to render the team
information.  We must first query the participant from the steps4impact
side.  This will identify the team that the participant is on.  Using
that, we can enumerate the members of the team and identify their IDs.
Then use that information to query information from Facebook to identify
the team members and retrieve their images.  Once all network
operations have quiesced, we can proceed to render the list.

It should be possible to render this using NSOperationQueue, which may
be simpler to follow than a Condition Variable and Semaphore used to
trigger the join of the network fan out.