models: adjust the model for Participant
The participant no longer has an `event_id` field.  This has been
replaced with an array of events (which may contain old events that they
have participated in).  Provide a helpful little `currentEvent` computed
property to get the current event that the participant is involved in.
This allows us to use the new backend.