Login: correct login flow for new user registration
Login: correct login flow for new user registration

The `getParticipant` end point will not create the user in the case that
the user does not already exist and returns an error if the user has not
been created.  If we fail to get the participant, we should create users
explicitly.  Adjust the login behaviour to match this expectation.

Co-authored By: Aalim Mulji <aalimmulji@gmail.com>